fix: fulltext may have Unix epoch as public date
The problem is due to that mercury may not be able to extract public
time from pages. Crawler didn't hand over vanilla rss article info to
the next stage where there probably be correct public time. We make it hand
over that info then in most cases the time will be correct.

Signed-off-by: oppilate <oppilate@tutanota.com>